Changed the 'git add <file>' to 'git add --all'
This change the git add behaviour so that not every file is added separtly to the git index. The async call crashed in our environment with '.git/index.lock - file exisits'. Seems to be a problem when calling git add in parallel from different git instances. So we call 'git add --all' to add all files and this is exactly what we want ;-)